Change default template data type from int64_t to long long

+++ cpp/include/ndim/dynamic_n_dim_diff_fenwick_tree.hpp
@@ -3,7 +3,7 @@
 
 namespace ndim {
 
-template <typename T = int64_t>
+template <typename T = long long>
 class DynamicNDimDiffFenwickTree {
 public:
     DynamicNDimDiffFenwickTree(const vector<int>& dims) : n_(dims.size()), tree_(dims), dims_(dims) {}
